only one flight event router now
